requirements.txt
aliyun-python-sdk-core==2.13.33
aliyun-python-sdk-core-v3==2.13.32
aliyun-python-sdk-kms==2.14.0
asgiref==3.3.4
beautifulsoup4==4.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==35.0.0
Django==3.2.16
django-cors-headers==3.2.0
django-ratelimit==2.0.0
django-redis==4.11.0
djangorestframework==3.11.0
drf-yasg==1.17.0
Faker==4.0.0
gunicorn==20.0.4
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.4.2
MarkupSafe==1.1.1
optionaldict==0.1.2
oss2==2.9.1
packaging==20.9
Pillow==8.3.2
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
pytz==2021.1
qiniu==7.2.8
redis==3.5.3
requests==2.22.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
sentry-sdk==0.14.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
wechatpy==1.8.6
xmltodict==0.12.0
